- Implemented basic temporal laplacian pyramid
This was to see what would happen if you visualize an amplified laplacian temporal level.
So far nothing interesting to see, but also a lot of calculations needed so large pyramids cannot keep up with a reasonable real-time framerate. Maybe temporal smoothing can help, but it is already so slow so will move to another algorithm.